requirements.txt
RapidFuzz==3.14.3
aiofiles==25.1.0
aiosqlite==0.22.1
beautifulsoup4==4.14.3
ddgs==9.11.4
fastapi==0.135.2
google-auth-oauthlib==1.3.0
google-auth==2.49.1
gspread==6.2.1
httpx==0.28.1
langgraph-checkpoint-sqlite==3.0.3
langgraph-checkpoint==4.0.1
langgraph-prebuilt==1.0.8
langgraph-sdk==0.3.12
langgraph==1.1.3
openai==2.30.0
pydantic-settings==2.13.1
pydantic==2.12.5
pydantic_core==2.41.5
python-dotenv==1.2.2
rich==14.3.3
structlog==25.5.0
tavily-python==0.7.23
tenacity==9.1.4
typer==0.24.1
uvicorn==0.42.0
websockets==16.0
